+++ src/api/server.d.ts
@@ -1,6 +1,0 @@
-/**
- * LiquidGov REST API Server
- */
-declare const app: import("express-serve-static-core").Express;
-export default app;
-//# sourceMappingURL=server.d.ts.map