+++ src/cli/simulate.d.ts
@@ -1,2 +1,0 @@
-export {};
-//# sourceMappingURL=simulate.d.ts.map