+++ tests/crowdfunding.test.d.ts
@@ -1,2 +1,0 @@
-export {};
-//# sourceMappingURL=crowdfunding.test.d.ts.map