+++ tests/delegation.test.d.ts
@@ -1,2 +1,0 @@
-export {};
-//# sourceMappingURL=delegation.test.d.ts.map