+++ tests/qv.test.d.ts
@@ -1,2 +1,0 @@
-export {};
-//# sourceMappingURL=qv.test.d.ts.map